Update the camera demo application
Replaced the basic Python web server with a custom one that can handle
asynchronous requests.

Added controls to the video page that allow the user to start and stop
the stream, and made the video playback more robust.

Added a gallery and controls to the snapshot page, and improved the
snapshot scripts. The first couple of images are now discarded, which
helps to ensure that the ISP has "warmed up" before the final image
is taken.